OPENEUROPA-2668: Move media steps into MediaCreationContext.
