+++ tests/Behat/FeatureContext.php
@@ -298,78 +298,4 @@ class FeatureContext extends RawDrupalContext {
     $this->assertVisuallyVisible($node);
   }
 
-  /**
-   * Creates media documents with the specified file names.
-   *
-   * // phpcs:disable
-   * @Given the following documents:
-   * | name 1 | file 1 |
-   * | name 2 | file 1 |
-   * |   ...  |   ...  |
-   * // phpcs:enable
-   */
-  public function createMediaDocuments(TableNode $file_table): void {
-    // Retrieve the url table from the test scenario.
-    $files = $file_table->getRows();
-
-    foreach ($files as $file_properties) {
-      $file = file_save_data(file_get_contents(drupal_get_path('module', 'oe_content') . '/tests/fixtures/' . $file_properties[1]), 'public://' . $file_properties[1]);
-      $file->setPermanent();
-      $file->save();
-
-      $media = \Drupal::service('entity_type.manager')
-        ->getStorage('media')->create([
-          'bundle' => 'document',
-          'name' => $file_properties[0],
-          'oe_media_file' => [
-            'target_id' => (int) $file->id(),
-          ],
-          'uid' => 0,
-          'status' => 1,
-        ]);
-
-      $media->save();
-    }
-  }
-
-  /**
-   * Creates media AVPortal photos with the specified URLs.
-   *
-   * // phpcs:disable
-   * @Given the following AV Portal photos:
-   * | url 1 |
-   * | url 2 |
-   * |  ...  |
-   * // phpcs:enable
-   */
-  public function createMediaAvPortalPhotos(TableNode $url_table): void {
-    // Retrieve the url table from the test scenario and flatten it.
-    $urls = $url_table->getRows();
-    array_walk($urls, function (&$value) {
-      $value = reset($value);
-    });
-
-    $pattern = '@audiovisual\.ec\.europa\.eu/(.*)/photo/(P\-.*\~2F.*)@i';
-
-    foreach ($urls as $url) {
-      preg_match_all($pattern, $url, $matches);
-      if (empty($matches)) {
-        continue;
-      }
-
-      // Converts the slash in the photo id.
-      $photo_id = str_replace("~2F", "/", $matches[2][0]);
-
-      $media = \Drupal::service('entity_type.manager')
-        ->getStorage('media')->create([
-          'bundle' => 'av_portal_photo',
-          'oe_media_avportal_photo' => $photo_id,
-          'uid' => 0,
-          'status' => 1,
-        ]);
-
-      $media->save();
-    }
-  }
-
 }
