+++ tests/features/publication.feature
@@ -4,7 +4,7 @@ Feature: Publication content creation
   As an editor
   I need to be able to create and see publication items
 
-  @cleanup:media @cleanup:file @run
+  @cleanup:media @cleanup:file
   Scenario: Creation of a Publication content through the UI.
     Given I am logged in as a user with the "create oe_publication content, access content, edit own oe_publication content, view published skos concept entities" permission
     # Create a "Document".
